The multiplication call fix provider for multiplication assignment now partially works. :boom: Roslyn y u no respect formatting trivia?

+++ NINNES.Platform/Shims/NESMath.cs
@@ -2,7 +2,7 @@
 
 namespace NINNES.Platform.Shims {
   public static class NESMath {
-    public static float Multiply(int a, int b) {
+    public static int Multiply(int a, int b) {
       throw new NotImplementedException();
     }
   }
